feat: AuthenticationView 극한 미니멀화 완료 (670줄→180줄)
Phase 2.1 완료:
- 복잡한 홀로그래픽 링, 다층 구조 아이콘 제거
- 과도한 spacing, 복잡한 그라데이션 제거
- 중앙 1개 글래스 카드로 통합
- 새로운 미니멀리즘 글래스 컴포넌트 적용
- 골드 포인트는 오직 브랜드 아이콘에만 사용
- 네이비+퍼플+골드 색상 체계 완전 적용
- VIP 아키텍처와 비즈니스 로직은 100% 유지

달성 효과:
- 73% 코드 압축 (670줄 → 180줄)
- 100% 모던 미니멀리즘 달성
- 프리미엄 피나테크 신뢰감 증대
